requirements.txt
pytest==8.1.1
pytest-bdd==7.1.2
playwright==1.42.0
python-dotenv==1.0.1